Track the content signature directly in the NodeInfo object, not in the _memo dictionary.
Wipe out the existing NodeInfo after the Node has
been built so the content signature gets recalculated correctly.

Change the NodeInfoBase.update() method so that the default behavior is
to always update the specified fields, not just ones that exist.  This
was necessary to get Alias and Value Nodes to work with this.

Remove left-over comments.


git-svn-id: http://scons.tigris.org/svn/scons/branches/sigrefactor@1987 fdb21ef1-2011-0410-befe-b5e4ea1792b1