Avoid generating content signatures unless they're actually used:
Don't automatically create a content signature after a File Node is
built/visited, and delay the .sconsign merge of the information until
just before the information will be written.  This means we'll only write
out information that's actually fetched and used as part of deciding
whether or not to rebuild any targets.


git-svn-id: http://scons.tigris.org/svn/scons/branches/sigrefactor@2003 fdb21ef1-2011-0410-befe-b5e4ea1792b1