- Add LIBOBJS support:
Added keyword parameter `add_libobj' to conf.CheckFunc.  It should be name of
library source file, a Node of this file, or a True value; in the last case,
function name followed by tested language's suffix will be the source file
name.  Object produced by compiling this source will be always added to
ALL_LIBOBJS construction variable, which is later used by Project to find
source files to distribute.  When function is not found, it will be also added
to LIBOBJS construction variable, whose contents are appended to LIBS on
finishing Configure context.  Both ALL_LIBOBJS and LIBOBJS are cleared on
Configure initialization.


git-svn-id: http://scons.tigris.org/svn/scons/branches/automake_model@2275 fdb21ef1-2011-0410-befe-b5e4ea1792b1