cell_contents not defined in python < 2.5. [Issue 2035]
Some versions of Python supports lexical scoping of local variables
(aka closures) but not the cell_contents attribute that would allow
the FunctionAction signature that take the content of these closure
cells into account. The access to the cell_contents attribute is now
protected with a try/except AttributeError so that at least the other
parts of the function signature are properly computed.

Benoit



git-svn-id: http://scons.tigris.org/svn/scons/branches/core@2936 fdb21ef1-2011-0410-befe-b5e4ea1792b1