Testing on a Mac turned up a problem that did not show up on linux. SK and I agree that we are surprised by that.
Both bld = env['BUILDERS']['PDF'] and bld = env['BUILDERS']['DVI'] had
lines like
bld.add_action('.tex', LaTeXAuxAction)  # for DVI
bld.add_action('.tex', PDFLaTeXAuxAction) # for PDF
run from the initialization of two tools. This was a BAD thing.

That cured some test failures. For the rest I made all the TEX tests 
initialize the environment with the os.environ PATH since the check 
for the existence of the tools with where_is that looks on the system path.

So the tests still quietly exit if the tool is not on the system path 
and now find if it is in a non-standard location like I have.



git-svn-id: http://scons.tigris.org/svn/scons/trunk@3599 fdb21ef1-2011-0410-befe-b5e4ea1792b1