Capture changes that seem to work (and pass all tests) to speed up long CPPPATH lists by 50% or more.  Needs clean up:  a lot of duplicated code, random comments, etc.
The gist of this is moving the logic from Node.get_implicit_deps()
into the Scanner proper so it do more efficient things with
a PathList-like object that can search for .h files in O(1)
not O(n).  Next steps should be trying to fold the cut-and-paste
separate Scanner class into the Base class, eliminating unnecessary
calls and cleaning up the Node call tree, etc.


git-svn-id: http://scons.tigris.org/svn/scons/branches/sgk_PathList@3790 fdb21ef1-2011-0410-befe-b5e4ea1792b1