Remove the feature that would build .pdf graphics files from .eps files for the pdf latex builder
That is if the .tex file has "\includegraphics{figure1}"
and the file figure1.eps then when using the .DVI builder
latex will find the file and all is fine.
However, when using the .PDF builder pdflatex can not 
process .eps files and will fail.

After this patch the user will need to add 
env.PDF('figure1.eps')

Update two tests that used the old feature and would fail otherwise
I could not come up with a way to test for a feature
that is removed. That is, I can write a test that works
before the update and fails after but not the other way around.



git-svn-id: http://scons.tigris.org/svn/scons/trunk@3882 fdb21ef1-2011-0410-befe-b5e4ea1792b1