http://scons.tigris.org/issues/show_bug.cgi?id=2345
The 'buffer' fixer simply replaces 'buffer( ... )' with 'memoryview( ... )',
which is incorrect for our cases, so these changes had to be done by hand and
a forward-compatibility class added.

The 'xrange' fixer was applied.  Manual changes were minimal: a few case in
test strings and one use of 'range' as an identifer in the same scope as
where 'xrange' was converted to 'range'.

The "sets15" compat function, which provided backward compatibility for Python
versions prior to 2.2, was removed as no longer needed.


git-svn-id: http://scons.tigris.org/svn/scons/trunk@4735 fdb21ef1-2011-0410-befe-b5e4ea1792b1