Start the deprecation cycle for the BuildDir() method and the build_dir keyword parameter.
Several existing tests were still using BuildDir() or build_dir; they were
converted to use VariantDir() and variant_dir.

New tests were added to validate that the --warn=deprecated-build-dir option
and the SetOption method did the right thing.  This led to the discovery that
a commonly-used test pattern provided by the infrastructure gobbled up too
much, causing tests to succeed when they should have failed.  Fixing the
pattern led to other tests needing to be fixed.

In the process, it was discovered that the SCONSFLAG environment variable was
not getting correctly reset to its original value.  Fixing this also caused
additional tests to misbehave, requiring them to be updated.

And test/Sig.py, which tests the deprecated SCons.Sig module, was moved to
the test/Deprecated directory.

All in all, quite a lot of action for what was supposed to be a simple change.


git-svn-id: http://scons.tigris.org/svn/scons/trunk@4920 fdb21ef1-2011-0410-befe-b5e4ea1792b1