Initial drop of bin/update-release-info.py.
This program automates the process of inserting the release configuration
information that is currently done by hand.  The configuration data now
all lives in a file, 'ReleaseConfig'.  Change the configuration file, run
the program, and, hey presto, the files are updated.

More work needs to be done; the documentation doesn't quite match with the
program and the regression tests only cover the release flows, but it's
important to get this stuff out before the next release is published so
people can start using it.


git-svn-id: http://scons.tigris.org/svn/scons/trunk@4981 fdb21ef1-2011-0410-befe-b5e4ea1792b1