Fixed a bug in the regression test to use the right filenames for each platform; it was causing the test to fail on Windows.
Added some documentation to bin/update-release-info.py.

In my last log message, I forgot to mention that QMTest/TestRuntest.py was
tweaked to make it less specialized for testing runtest.py.  It can now be
used to test other programs; in this case, bin/update-release-info.py.  The
defaults are unchanged so that it will normally test runtest.py.


git-svn-id: http://scons.tigris.org/svn/scons/trunk@4987 fdb21ef1-2011-0410-befe-b5e4ea1792b1